Introduce source-aware chart processor provider registry
