+++ src/aviationapi/chart_processor/app/providers/base.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+
+class ChartProvider(ABC):
+    source = ""
+
+    @abstractmethod
+    def process_packet(self, packet, airac):
+        pass
