+++ src/aviationapi/chart_processor/app/providers/registry.py
@@ -1,0 +1,7 @@
+from aviationapi.chart_processor.app.providers.faa_tpp import FaaTppChartProvider
+
+PROVIDERS = {FaaTppChartProvider.source: FaaTppChartProvider()}
+
+
+def get_provider(source):
+    return PROVIDERS.get(source)
